src/core/performance/cpu/vectorization.py
"""CPU Vectorization Framework for Adaptive Attention Tiling.

This module provides vectorized operations optimized for CPU execution,
focusing on attention computation, pattern dynamics, and geometric flows.
"""

import numpy as np
import torch
import torch.nn.functional as F
import time
import psutil
from typing import List, Tuple, Optional, Union, Callable
from dataclasses import dataclass
from functools import wraps

@dataclass
class VectorizationMetrics:
    """Metrics for vectorized operations."""
    execution_time: float
    memory_usage: float
    vectorization_efficiency: float
    operation_type: str

class VectorizationOptimizer:
    """Optimizes operations using CPU vectorization techniques."""
    
    def __init__(self, 
                 enable_profiling: bool = True,
                 use_mixed_precision: bool = True,
                 chunk_size: int = 1024):
        self.enable_profiling = enable_profiling
        self.use_mixed_precision = use_mixed_precision
        self.chunk_size = chunk_size
        self.metrics: List[VectorizationMetrics] = []

    @staticmethod
    def profile_vectorization(func: Callable) -> Callable:
        """Decorator to profile vectorized operations."""
        @wraps(func)
        def wrapper(self, *args, **kwargs):
            if not self.enable_profiling:
                return func(self, *args, **kwargs)
            
            process = psutil.Process()
            start_memory = process.memory_info().rss / 1024**2  # MB
            start_time = time.perf_counter()
            
            result = func(self, *args, **kwargs)
            
            end_time = time.perf_counter()
            end_memory = process.memory_info().rss / 1024**2  # MB
            
            execution_time = (end_time - start_time) * 1000  # Convert to ms
            memory_usage = end_memory - start_memory
            
            # Estimate vectorization efficiency
            vectorization_efficiency = self._estimate_vectorization_efficiency(result)
            
            self.metrics.append(VectorizationMetrics(
                execution_time=execution_time,
                memory_usage=memory_usage,
                vectorization_efficiency=vectorization_efficiency,
                operation_type=func.__name__
            ))
            
            return result
        return wrapper

    @profile_vectorization
    def vectorize_attention(self, 
                          query: torch.Tensor,
                          key: torch.Tensor,
                          value: torch.Tensor) -> torch.Tensor:
        """Vectorized attention computation."""
        # Ensure contiguous memory layout for better CPU performance
        query, key, value = map(lambda x: x.contiguous(), (query, key, value))
        
        # Split into chunks for better cache utilization
        chunk_size = min(self.chunk_size, query.size(0))
        result = []
        
        for i in range(0, query.size(0), chunk_size):
            chunk_end = min(i + chunk_size, query.size(0))
            q_chunk = query[i:chunk_end]
            k_chunk = key[i:chunk_end]
            v_chunk = value[i:chunk_end]
            
            # Compute attention for chunk
            scores = torch.matmul(q_chunk, k_chunk.transpose(-2, -1))
            scores = scores / np.sqrt(k_chunk.size(-1))
            attn_weights = F.softmax(scores, dim=-1)
            chunk_result = torch.matmul(attn_weights, v_chunk)
            result.append(chunk_result)
        
        return torch.cat(result, dim=0)

    @profile_vectorization
    def vectorize_pattern_dynamics(self, 
                                 pattern: torch.Tensor,
                                 flow: torch.Tensor) -> torch.Tensor:
        """Vectorized pattern dynamics computation."""
        # Ensure contiguous memory layout
        pattern, flow = pattern.contiguous(), flow.contiguous()
        
        # Compute evolution rate based on pattern characteristics
        evolution_rate = self._compute_evolution_rate(pattern)
        
        # Apply flow to pattern using efficient operations
        updated_pattern = pattern + evolution_rate * flow
        return F.normalize(updated_pattern, p=2, dim=-1)

    @profile_vectorization
    def vectorize_geometric_flow(self,
                               metric: torch.Tensor,
                               connection: torch.Tensor) -> torch.Tensor:
        """Vectorized geometric flow computation."""
        # Ensure contiguous memory layout
        metric, connection = metric.contiguous(), connection.contiguous()
        
        # Compute Christoffel symbols efficiently
        christoffel = torch.einsum('...ij,...jkl->...ikl', metric, connection)
        
        # Compute flow components
        flow = torch.zeros_like(metric)
        for i in range(metric.size(-1)):
            for j in range(metric.size(-1)):
                flow[..., i, j] = torch.sum(christoffel[..., i, :, j], dim=-1)
        
        return flow

    def _compute_evolution_rate(self, pattern: torch.Tensor) -> torch.Tensor:
        """Compute evolution rate based on pattern characteristics."""
        # Compute pattern statistics
        mean = torch.mean(pattern, dim=-1, keepdim=True)
        std = torch.std(pattern, dim=-1, keepdim=True)
        
        # Adaptive rate based on pattern variability
        base_rate = 0.01
        return base_rate * (1 + torch.tanh(std / mean))

    def _estimate_vectorization_efficiency(self, result: torch.Tensor) -> float:
        """Estimate vectorization efficiency based on memory layout and operations."""
        # Check memory contiguity
        contiguous_score = 1.0 if result.is_contiguous() else 0.5
        
        # Check alignment
        alignment_score = 1.0 if result.storage_offset() % 16 == 0 else 0.7
        
        # Check size factors
        size_score = 1.0 if all(s % 16 == 0 for s in result.shape) else 0.8
        
        return (contiguous_score + alignment_score + size_score) / 3

    def get_metrics(self) -> List[VectorizationMetrics]:
        """Get collected vectorization metrics."""
        return self.metrics

    def clear_metrics(self) -> None:
        """Clear collected metrics."""
        self.metrics.clear()
